scs/requirements.txt
python-vlc
gTTS
Pillow